pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "limenex"
version = "0.1.0"
description = "Agentic AI execution governance layer"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.black]
target-version = ["py310", "py311", "py312"]
